--- a/.travis.yml
+++ b/.travis.yml
@@ -14,6 +14,7 @@ script:
 
 matrix:
   allow_failures:
+    - rust: beta
     - rust: nightly
 
 env: