--- a/install_deps.sh
+++ b/install_deps.sh
@@ -3,6 +3,12 @@
 set -eux
 
 BUILD_DIR=$PWD
+MAKE_FLAGS=""
+
+# Currently only Linux is supported.
+case "$(uname)" in
+    "Linux") MAKE_FLAGS="-j$(nproc)";;
+esac
 
 #
 # gf-complete
@@ -12,7 +18,7 @@ cd gf-complete/
 git checkout a6862d1
 ./autogen.sh
 ./configure --disable-shared --with-pic --prefix $BUILD_DIR
-make install
+make $MAKE_FLAGS install
 cd ../
 
 #
@@ -23,7 +29,7 @@ cd jerasure/
 git checkout de1739c
 autoreconf --force --install
 CFLAGS="-I${BUILD_DIR}/include" LDFLAGS="-L${BUILD_DIR}/lib" ./configure --disable-shared --enable-static --with-pic --prefix $BUILD_DIR
-make install
+make $MAKE_FLAGS install
 cd ../
 
 #
@@ -39,4 +45,4 @@ if [ "$(uname)" == "Darwin" ]; then
 fi
 CFLAGS=$CFLAGS LIBS="-lJerasure" LDFLAGS="-L${BUILD_DIR}/lib" ./configure --disable-shared --with-pic --prefix $BUILD_DIR
 patch -p1 < ../liberasurecode.patch # Applies a patch for building static library
-make install
+make $MAKE_FLAGS install