Make build a bit faster using `-j` option.

# Before

```console
$ time cargo build
   Compiling liberasurecode v1.0.2
   Compiling libc v0.2.43
    Finished dev [unoptimized + debuginfo] target(s) in 1m 19s
cargo build  62.55s user 12.33s system 94% cpu 1:19.38 total
```

# After

```console
$ time cargo build
   Compiling liberasurecode v1.0.2
   Compiling libc v0.2.43
    Finished dev [unoptimized + debuginfo] target(s) in 1m 05s
cargo build  61.67s user 11.71s system 112% cpu 1:05.20 total
```