@sile 2018-11-06
I would be glad if you would add `allow_failures: [{rust: beta}]` line to `.travis.yml` file for suppressing non essential failures of the CI (cf. https://github.com/frugalos/cannyls/pull/8/files#diff-354f30a63fb0907d4ad57269548329e3).