@yuezato 2018-11-06
> Can somebody try this change on macOS?

I tried this PR and succeeded in `cargo build`.
For MacOS (Darwin), I'd like to add the following line to `install_deps.sh`:
```sh
case "$(uname)" in
    "Linux") MAKE_FLAGS="-j$(nproc)";;
    "Darwin") MAKE_FLAGS="-j$(sysctl -n hw.ncpu)"
esac
```